# Parquet multi kernel optimization example.