@nvdbaranec 2023-06-26
Open questions:
- We would need to figure out what the right subset of kernels actually is.  Based on our feature set, it's easy to imagine 16+ kernels with slightly different features.  But that might be overkill.  I tend to think the primary way of breaking them down should be based on shared memory usage differences.

- There's more optimization that could be pushed through here.  For example, the individual "copy this piece of data" functions are shared between the new and old kernels, but they all do a dictionary if-check.  This could be `if constexpr`'d away if we wanted to go crazier. 

- This file layout is just a suggestion.  `page_data.cu` was getting pretty crazy so I wanted to try out some ideas on splitting things up.