@nvdbaranec 2023-06-26
Also, there will be a decent amount of work to get this merged with the big string decoding change from @etseidl