@GregoryKimball 2024-01-29
FYI adding specialized decode kernels to parquet reader came up again today in a discussion about the query performance roadmap. I believe the next step was to group the different data types, nullability and encodings by register usage. Then from there we could start adding specialized kernels one at a time. 

Please let me know if you agree.